requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.18.0
selenium>=4.16.0
schedule>=1.2.0
python-dotenv>=1.0.0
